fix: keep overview projection toggle mounted during reload
